вынес фильтрацию объявлений в отдельный модуль
